ImageProcessing split into separate files
