+++ Sources/Core/Processing/ImageProcessors+Anonymous.swift
@@ -1,0 +1,26 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2015-2021 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+extension ImageProcessors {
+    /// Processed an image using a specified closure.
+    public struct Anonymous: ImageProcessing, CustomStringConvertible {
+        public let identifier: String
+        private let closure: (PlatformImage) -> PlatformImage?
+
+        public init(id: String, _ closure: @escaping (PlatformImage) -> PlatformImage?) {
+            self.identifier = id
+            self.closure = closure
+        }
+
+        public func process(_ image: PlatformImage) -> PlatformImage? {
+            self.closure(image)
+        }
+
+        public var description: String {
+            "AnonymousProcessor(identifier: \(identifier)"
+        }
+    }
+}
