+++ Sources/Core/Processing/ImageProcessors.swift
@@ -1,0 +1,8 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2015-2021 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+/// A namespace for all processors that implement `ImageProcessing` protocol.
+public enum ImageProcessors {}
